hospitalmanagementsystem/requirements.txt
amqp
asgiref
attrs
billiard
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
cryptography
defusedxml
Django
django-encrypted-model-fields
django-fernet-fields
django-redis
django-simple-history
django_celery_results
djangorestframework
djangorestframework_simplejwt
djoser
drf-spectacular
gunicorn
idna
inflection
jsonschema
jsonschema-specifications
kombu
oauthlib
packaging
prompt_toolkit
psycopg2-binary
pycparser
PyJWT
python-dateutil
python-decouple
python3-openid
PyYAML
redis
referencing
requests
requests-oauthlib
rpds-py
six
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth